Домашнее задание. Вывести все чётные числа до числа N.

+++ Seminar001Task8/Program.cs
@@ -1,6 +1,6 @@
 ﻿Console.WriteLine("Введите число:");
-double N = double.Parse(Console.ReadLine()??"0");
-for (double i = 2; i <= N;i = i + 2)
+int N = int.Parse(Console.ReadLine()??"0");
+for (int i = 2; i <= N;i = i + 2)
 {
     Console.Write($"{i}, ");
 }
